refactor(server): extract /api/campaign/* into a router module
Fourth route-group extraction (publishing deferred — it's the most
entangled). Moves all 9 campaign handlers + the campaign-only
enrichAngleForCampaign helper into src/server/routes/campaign.js, mounted
`app.use('/api/campaign', campaignRouter)`.

Auth: MIXED group — GET /api/campaign/:id is public, the other 8 are
requireAuth — so auth stays PER-ROUTE (no mount-level requireAuth), per
the convention set for mixed groups. requireAuth is imported by the module.

Routes (paths byte-identical, guard-verified): :id/archive, :id/unarchive,
arcs/:brandProfileId, plan, reset/:id, create, list/:brandProfileId, :id,
generate/:id.

Also extracts a SHARED helper the dep scan surfaced:
- src/server/content-table.js — ensureGeneratedContentTable (idempotent
  per-brand generated_content_<id> table create/upgrade). Shared by the
  content-generator route (still in server.js) AND campaign generate, so
  it became a shared module both import — same pattern as streams.js.

enrichAngleForCampaign (campaign-only) moved into the router module;
generateArticleImage is an inner closure in the generate handler (moved
with it). Deps otherwise all already extracted: pool, anthropic,
extractJSON/safeParseLLM, requireAuth/verifyBrandAccess,
stripScaffoldingArtifacts, buildImagePrompt/generateHeroImage + fs/path.

Test plan:
- node --check server.js + campaign.js + content-table.js -> OK
- npm run lint (no-undef) -> clean (caught fs/path; added them)
- route-inventory guard -> 213 byte-identical (9 campaign paths)
- vitest -> 93/93

Process note: the campaign generate + enrich helper both contain prompt
template literals with `}` at column 0, which broke a naive "first ^}"
function-end detection mid-build (caught by node --check). Re-cut using
boundary-by-next-statement detection. No bad code shipped.

Rollback: revert; verbatim move + 2 small shared/router modules.
